Add config of light transition

+++ src/Extensions/LightExtensionMethods/LightParameters.cs
@@ -5,7 +5,7 @@ namespace NetEntityAutomation.Extensions.LightExtensionMethods;
 public record LightParameters
 {
     [JsonPropertyName("transition")]
-    public long? Transition { get; init; }
+    public double? Transition { get; init; }
     
     [JsonPropertyName("brightness_pct")]
     public long? BrightnessPct { get; init; }
